third_party/blink/web_tests/platform/generic/storage/indexeddb/mozilla/event-source-expected.txt

Test IndexedDB's event.target.source in success callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "event-source.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
source = event.target.source;
PASS source is null
objectStore = db.createObjectStore(objectStoreName, { autoIncrement: true });
request = objectStore.add({});
source = event.target.source;
PASS source is objectStore
PASS successfullyParsed is true

TEST COMPLETE
